requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
azure-common==1.1.28
azure-core==1.35.1
azure-identity==1.25.0
azure-keyvault-secrets==4.10.0
azure-search==1.0.0b2
azure-search-documents==11.7.0b1
beautifulsoup4==4.13.5
blinker==1.9.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
cryptography==45.0.6
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-appconf==1.1.0
django-cors-headers==4.4.0
django-filter==25.1
django-imagekit==5.0.0
django-rest-auth==0.9.5
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
docopt==0.6.2
dotenv==0.9.9
drf-extensions==0.8.0
fcm-django==2.3.1
fiona==1.10.1
firebase_admin==7.1.0
Flask==3.1.2
flask-cors==6.0.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
Js2Py==0.74
Markdown==3.8.2
MarkupSafe==3.0.2
msal==1.33.0
msal-extensions==1.3.1
msgpack==1.1.1
msrest==0.7.1
nltk==3.9.1
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
openai==1.107.3
packaging==25.0
pandas==2.3.2
pilkit==3.0
pillow==11.3.0
pipwin==0.5.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg==3.2.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
pyjsparser==2.7.1
PyJWT==2.10.1
pyparsing==3.2.3
PyPDF2==3.0.1
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
regex==2025.9.1
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
swapper==1.4.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uv==0.8.16
Werkzeug==3.1.3
